==> 0A.data_check_n_clean.ipynb <==
import sys
import pandas as pd
import os
import seaborn as sns
import xarray as xr
import numpy as np
from sklearn.linear_model import LinearRegression
import matplotlib.pyplot as plt
import datetime
from datetime import datetime
from datetime import date
import re
import netCDF4 as nc
from scipy import stats

# open all the files 

import glob
mylist = [f for f in glob.glob("/Users/caropark/Dropbox (Harvard University)/DataByCountry/*")]
mylist.remove('/Users/caropark/Dropbox (Harvard University)/DataByCountry/Icon\r')

start = '/Users/caropark/Dropbox (Harvard University)/DataByCountry/'
end_fao = '/fao_data.nc'
end_cal = '/crop_calendar_data_sage.nc'
end_sif = '/csif_data.nc'
end_harvarea = "/harvest_area_fraction_data_sage_all.nc"

myends = ["end_cal", "end_sif", "end_harvarea"]      

remove=[]

for i in mylist: 
    for end in myends:        
        if os.path.isfile(str(i) + eval(end)):
            pass
        else:         #            print(end, i)        
            remove.append(i)
            remove = [*set(remove)]

mylist_cleaned = set(mylist).difference(set(remove))

print("safely removed countries w.o fao/sif/mask data:",  len(mylist_cleaned) + len(remove)== len(mylist), "\n")

remove = []

for i in mylist_cleaned:
    harvarea = xr.open_dataset(str(i)+end_harvarea)['harvest_area_frac_nearest']
    counts = harvarea.where(harvarea>0).groupby('crop').count(...)
    counts = counts.where(counts>0, drop=True)['crop'].values
    
    calendar = xr.open_dataset(str(i)+end_cal)['plant']
    
    if counts.size==0 or np.sum(calendar.values)==0:
        print("List is empty:", i)
        remove.append(i)
        remove = [*set(remove)]

allcountries = list(set(mylist_cleaned).difference(set(remove)))
print("output: allcountries")
